backend/requirements.txt
Flask
Flask-Cors
maafw == 5.2.3
pillow